requirements.txt
absl-py==2.3.0
astunparse==1.6.3
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.4
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.2.1
protobuf==5.29.5
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.4
rich==14.0.0
setuptools==80.9.0
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow_keras==0.1
termcolor==3.1.0
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
